test: add comprehensive test coverage for all modules
- Add extensive Bson type tests (IConvertible, DateTime, Decimal128, etc.)
- Add DocumentCollection edge case and async tests
- Add Core module tests (TransactionManager, LockManager, DataPageAccess)
- Add Query module tests (Parser, Executor, Pipeline, Optimizer)
- Add Serialization tests (AotBsonMapper, BsonSerializer, SizeCalculator)
- Add Storage tests (WriteAheadLog, DiskStream)
- Add Security and ForeignKey tests
- Total test count: 2569 tests passing

+++ TinyDb.Tests/Bson/BsonValueCoverageTests.cs
@@ -131,7 +131,7 @@ public class BsonValueCoverageTests
     public async Task BsonDecimal128_Coverage()
     {
         var val = new BsonDecimal128(123.45m);
-        await Assert.That(val.Value).IsEqualTo(123.45m);
+        await Assert.That((decimal)val.Value).IsEqualTo(123.45m);
         await Assert.That(val.ToString()).IsEqualTo("123.45");
         
         await Assert.That(val.ToBoolean(null)).IsTrue();
